Rename Fit to Fit_VLE, abstract GeneralFit
