+++ appPy/cli/fit.py
@@ -6,8 +6,8 @@ import click
 from src.utils.systems import validate_system_or_swap
 from src.utils.errors import CLI_error_boundary, AppException
 from src.utils.io.plot import pause_to_keep_charts
-from src.plot.Fit_plot import Fit_plot
-from src.fit.Fit import default_model, supported_model_names
+from src.plot.Fit_VLE_plot import Fit_VLE_plot
+from src.fit.Fit_VLE import default_model, supported_model_names
 
 model_list = ", ".join(supported_model_names)
 
@@ -27,7 +27,7 @@ def cli_fit(compound1, compound2, model, datasets, params, consts, xy, txy, gamm
     """Fit binary VLE data with a given model, datasets, and optionally with specified initial parameters."""
 
     compound1, compound2 = validate_system_or_swap(compound1, compound2)
-    fit = Fit_plot(compound1, compound2, model, datasets, parse_params(params), parse_consts(consts))
+    fit = Fit_VLE_plot(compound1, compound2, model, datasets, parse_params(params), parse_consts(consts))
     if not skip: fit.optimize()
     fit.report()
